Add files via upload
This little program created for Windows 11 merely switches the Internet on or off. No indicator included so it's literally a one click affair. Just look in your system tray to see if the internet is on or off. I did this because I kept getting an OK confirmation checkbox up and I didn't want this. Coded in ScITE. POnly works if your run it as admin. compile for 64bit version.